fixed note id bug and refactored component directories
