+++ src/__test__/dashboard.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { configure, mount } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import Dashboard from '../component/dashboard';
+import Dashboard from '../dashboard/dashboard';
 
 configure({ adapter: new Adapter() });
