+++ src/__test__/note-form.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { configure, mount } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import NoteForm from '../component/note-form';
+import NoteForm from '../note-form/note-form';
 
 configure({ adapter: new Adapter() });
