+++ src/main.js
@@ -2,7 +2,7 @@
 
 import React from 'react';
 import { render as reactDomRender } from 'react-dom';
-import App from '../src/component/app';
+import App from './app/app';
 
 const container = document.createElement('div');
 document.body.appendChild(container);
