+++ src/note-item/note-item.js
