+++ src/styles/_base.scss
@@ -1,0 +1,19 @@
+@import 'vars';
+
+body {
+    background: $background-color;
+    font-family: sans-serif;
+    color: $font-color;
+}
+
+a {
+    color: $font-color;
+}
+
+a:hover {
+    text-decoration: none;
+}
+
+.error {
+    color: $error;
+}
