+++ src/styles/_vars.scss
@@ -1,0 +1,13 @@
+//--------------------------------------------------------------
+// COLORS
+//--------------------------------------------------------------
+$background-color: #5f5eed;
+$font-color: rgba(17, 17, 17, 0.8);
+
+$modal-color: rgba(0,0,0,0.4);
+$modal-background: #FFF;
+$error: #ff0000;
+
+//--------------------------------------------------------------
+// SIZE
+//--------------------------------------------------------------
